ESLint/Travis/build issues
