+++ .gitignore
@@ -34,7 +34,6 @@ vendor/opal-compiler.js
 test/fixtures/*.backup
 /tmp
 /test-results.xml
-*.tmp
 
 # packed NPM packages
 /*.tgz
