+++ test/unit/transpile.js
@@ -3,9 +3,7 @@
 const expect = require('chai').expect
 const transpile = require('../../lib/transpile')
 const path = require('path')
-const tmp = require('tmp')
-const fs = require('fs')
-const exec = require('child_process').exec
+const alternateCompilerTest = require('../support/alternateCompilerTest')
 
 describe('transpile', function(){
   const wpContext = {
@@ -22,37 +20,6 @@ describe('transpile', function(){
     return transpile(code, targetOptions, wpContext).code
   }
 
-  function runTestAgainstOtherFile(code, compilerAbsolutePath, callback) {
-    tmp.file(function (err, tmpPath, fd, cleanup) {
-      fs.writeFile(tmpPath, code, function (err) {
-        if (err) {
-          cleanup()
-          return callback(err)
-        }
-
-        const environment = Object.assign({
-          OPAL_COMPILER_PATH: compilerAbsolutePath
-        }, process.env)
-        exec(`node ${tmpPath}`, {
-          env: environment
-        },
-        function (err, stdout) {
-          if (err) {
-            cleanup()
-            return callback(err)
-          }
-
-          try {
-            const result = stdout.trim()
-            return callback(null, result)
-          }
-          finally {
-            cleanup()
-          }
-        })
-      })
-    })
-  }
 
   it('compiles an endpoint', function() {
     var result = doTranspile('HELLO=123')
@@ -83,12 +50,10 @@ describe('transpile', function(){
   })
 
   it('loads an Opal compiler from a configurable file', function(done) {
-    const compilerAbsPath = path.resolve(__dirname, '../support/tweakedOpalCompiler.js')
-    const compilerRelativePath = './test/support/tweakedOpalCompiler.js'
-    const code = `var transpile = require('./lib/transpile');\n`+
-    `console.log(transpile('HELLO=123', {filename: '/foo.rb', relativeFileName: 'foo.rb'}, {path: 'the_loader_path'}).code);`
+    const code = `var transpile = require('lib/transpile');\n`+
+    "console.log(transpile('HELLO=123', {filename: '/foo.rb', relativeFileName: 'foo.rb'}, {path: 'the_loader_path'}).code);"
 
-    runTestAgainstOtherFile(code, compilerAbsPath, function(err, result) {
+    alternateCompilerTest.execute(code, function(err, result) {
       if (err) { return done(err) }
       expect(result).to.include('Generated by Opal 0.10.0.beta2.webpacktest')
       return done()
@@ -96,13 +61,10 @@ describe('transpile', function(){
   })
 
   it('passes custom configured Opal through', function(done) {
-    const compilerAbsPath = path.resolve(__dirname, '../support/tweakedOpalCompiler.js')
-    const compilerRelativePath = './test/support/tweakedOpalCompiler.js'
-    const code = `var transpile = require('./lib/transpile');\n`+
-    `const code = transpile('the code', {filename: '${compilerAbsPath}', relativeFilename: '${compilerRelativePath}'}, {path: 'the_loader_path'}).code\n` +
-    `console.log(code)`
+    const code = `var transpile = require('lib/transpile');\n`+
+    `const code = transpile('the code', {filename: '${alternateCompilerTest.absolutePath}', relativeFilename: '${alternateCompilerTest.relativePath}'}, {path: 'the_loader_path'}).code\n` + 'console.log(code)'
 
-    runTestAgainstOtherFile(code, compilerAbsPath, function(err, result) {
+    alternateCompilerTest.execute(code, function(err, result) {
       if (err) { return done(err) }
       expect(result).to.eq('process = undefined;\nthe code')
       return done()
